# smartlock-docs/enterprise/16_API_Spec.yaml
...
openapi: 3.1.0
info:
  title: Smart Lock AI Support & Service Dispatch SaaS — 同步介面契約
  version: "1.1"
  summary: 平台所有同步 HTTP / WebSocket 介面的單一契約入口
  description: |
    本檔涵蓋平台 6 子系統的同步介面：

    | 介面 | 提供者 → 消費者 |
    |---|---|
    | 品牌營運控制平面（工單/派工/結算/帳務/治理；runtime export 517 端點）| api → web(dispatch) / 平台 console |
    | agent 服務間旁路 `/internal/*`（4 端點）| agent → api |
    | LINE 入站 webhook `POST /callback` | LINE 平台 → agent（見 webhooks 區）|
    | OHS API（技師查詢/媒合/排班/認證）| technician-platform → 各品牌 api 消費 |
    | knowledge-refinery 審核佇列 API | refinery web → refinery svc（摘要，見 refinery tag）|

    **組織原則**：runtime 517 端點不逐一展開；以 tags + 各領域代表性資源組織，其餘端點見各
    tag `externalDocs` 指向的系統契約文件（完整端點總表 `../api/P2/06` §3）。

    **通用約定**（`../api/P2/06` §1）：
    - 版本切分：v1（legacy）掛 `/api/v1` 前綴且所有回應帶 `Deprecation: true` header；
      v2（tenant-scoped）走 `/tenants/{tenantId}/...` flat path。
    - 認證：browser HTTP 走本站 proxy + HttpOnly access/refresh cookie；CLI/mobile 可用
      Casdoor OIDC bearer。服務間優先 `X-Service-Credential`，`X-Internal-Token` 只作
      CR-0190 migration fallback。
    - 回應信封：成功 `{data, error: null}`（ApiResponse）/ 游標分頁 `CursorPage`；
      錯誤統一 RFC7807 superset（`application/problem+json`，見 Problem schema）。
    - 冪等：寫入類支援 `Idempotency-Key` header，TTL 24h。
    - SoD：雙簽/三簽端點需 `X-Initiator` / `X-Approver`（+ 選填 `X-Executor`），
      任二相同 → 403 SOD_VIOLATION。
    - 即時推播：WS 頻道契約見 `x-websocket-channels`；fanout 傳輸設計見 ./17_AsyncAPI.yaml。
...
      內部 MCP 協議，非 HTTP，不進本契約）。
  contact: { name: 平台架構組 }
...
servers:
...
security:
  - casdoorOIDC: []
...
paths:
...
  /api/v2/platform/service-principals:
    get:
      tags: [service-identity]
      operationId: listServicePrincipals
      summary: 列出 service principal metadata；不回 credential secret/hash
      responses:
        "200": { description: Principal 清單, content: { application/json: { schema: { $ref: "#/components/schemas/ApiResponse" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [service-identity]
      operationId: createServicePrincipal
      summary: 建立 principal 並一次性簽發 hash-only-at-rest credential
      parameters: [{ $ref: "#/components/parameters/idempotencyKey" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ServicePrincipalCreateBody" } } }
      responses:
        "201": { description: plaintext credential 僅此 response 可見, content: { application/json: { schema: { $ref: "#/components/schemas/IssuedServiceCredentialEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v2/platform/service-credentials/{credentialId}:rotate:
    post:
      tags: [service-identity]
      operationId: rotateServiceCredential
      summary: 輪替 credential（0～86400 秒 overlap）
      parameters:
        - { name: credentialId, in: path, required: true, schema: { type: string } }
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ServiceCredentialRotateBody" } } }
      responses:
        "200": { description: 新 plaintext credential 僅此 response 可見, content: { application/json: { schema: { $ref: "#/components/schemas/IssuedServiceCredentialEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v2/platform/service-credentials/{credentialId}:revoke:
    post:
      tags: [service-identity]
      operationId: revokeServiceCredential
      summary: 立即撤銷 credential
      parameters:
        - { name: credentialId, in: path, required: true, schema: { type: string } }
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ServiceCredentialRevokeBody" } } }
      responses:
        "200": { description: 已撤銷 }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    casdoorOIDC:
      type: openIdConnect
      openIdConnectUrl: https://casdoor.platform.internal/.well-known/openid-configuration  # [待確認] 正式 Casdoor URL
      description: |
        Casdoor 統一 IdP（ADR-P003）。各服務驗 OIDC bearer token；角色由 role claim
        提供，四方 RBAC（Super Admin / 租戶 Admin / 派工小編 / 技師跨租戶身分）
        deny-by-default enforce（ADR-P006）。守衛鏈：require_tenant（X-Tenant-ID 與
        claim 一致）→ role_required(*roles) → require_platform_admin（跨品牌視角）
        → require_sod_actors / require_keeper_role（特權操作）。
        系統角色 12 個：admin / reviewer / technician / brand_oem / line_user /
        accounting / supervisor / dispatcher / customer_service / auditor /
        family_reviewer / distributor。
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: CR-0190 migration fallback only；production usage 須連續一個 release window 為零後移除。
    serviceCredential:
      type: apiKey
      in: header
      name: X-Service-Credential
      description: Per-workload opaque credential；hash-only、audience/scope/tenant/expiry/revoke/audit。
    ohsServiceCredential:
      type: apiKey
      in: header
      name: X-Service-Credential
      description: "品牌 api → technician-platform OHS 現行可用的受控 opaque credential；OD-001 仍決定最終是否改 OIDC client-credentials。"
    wsCookieAuth:
      type: apiKey
      in: cookie
      name: smartlock_access_token
      description: WS/SSE 認證只走 HttpOnly cookie；URL 不得含 access/refresh token。跨 host 無共享 cookie 時 realtime disabled。
  parameters:
...
    idempotencyKey:
      { name: Idempotency-Key, in: header, required: false, schema: { type: string }, description: "寫入類冪等鍵，TTL 24h；重放回原結果" }
...
  responses:
...
    Forbidden:
      description: 授權失敗（FORBIDDEN、ACCOUNT_DISABLED、TENANT_MISMATCH、SOD_VIOLATION、KEEPER_FORBIDDEN）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: 資源不存在（NOT_FOUND）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Conflict:
      description: 狀態衝突 / 冪等重放（CONFLICT）
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  schemas:
...
    ServicePrincipalCreateBody:
      type: object
      required: [name, audiences, scopes, credential_expires_at]
      properties:
        name: { type: string }
        audiences: { type: array, items: { type: string } }
        scopes: { type: array, items: { type: string } }
        allowed_tenant_ids: { type: array, items: { type: string } }
        allow_all_tenants: { type: boolean, default: false }
        credential_expires_at: { type: string, format: date-time }
    ServiceCredentialRotateBody:
      type: object
      required: [expires_at]
      properties:
        expires_at: { type: string, format: date-time }
        overlap_seconds: { type: integer, minimum: 0, maximum: 86400, default: 0 }
    ServiceCredentialRevokeBody:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    IssuedServiceCredentialEnvelope:
      type: object
      description: plaintext `credential` writeOnly 且只在 create/rotate 當下回傳。
      properties:
        data:
          type: object
          properties:
            principal_id: { type: string }
            credential_id: { type: string }
            credential: { type: string, writeOnly: true }
            credential_prefix: { type: string }
            expires_at: { type: string, format: date-time }
            secret_visible_once: { type: boolean, const: true }
    Problem:
      type: object
      description: "RFC7807 superset 錯誤信封；type 為 urn:smartlock:error:{error_code_lowercase}（識別字串非可抓取 URL）；error_code/message/request_id 為 extension members"
      required: [type, title, status, error_code, request_id]
      properties:
        type: { type: string, examples: ["urn:smartlock:error:validation_error"] }
        title: { type: string, examples: ["Validation Error"] }
        status: { type: integer, examples: [422] }
        detail: { type: string }
        instance: { type: string }
        error_code: { type: string, examples: [VALIDATION_ERROR] }
        message: { type: string }
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string, examples: [body.amount] }
              issue: { type: string }
              type: { type: string }
    ApiResponse:
      type: object
      description: 成功回應信封
      properties:
        data: {}
        error: { type: "null" }
...